VE model code
