+++ test_fixtures/vilbert_ve/experiment_from_huggingface.jsonnet
@@ -31,11 +31,11 @@ local model_name = "epwalsh/bert-xsmall-dummy";
     "image_intermediate_size": 50,
     "image_attention_dropout": 0.0,
     "image_hidden_dropout": 0.0,
-    "v_biattention_id": [0, 1],
-    "fixed_v_layer": 0,
+    "image_biattention_id": [0, 1],
+    "image_fixed_layer": 0,
 
-    "t_biattention_id": [0, 1],
-    "fixed_t_layer": 0,
+    "text_biattention_id": [0, 1],
+    "text_fixed_layer": 0,
 
     "combined_hidden_size": 200,
     "combined_num_attention_heads": 2,
